refactor: SseChannel 추상화 및 inmemory 패키지 분리
- InMemorySseChannel 구현
